requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
python-dateutil>=2.8.2
Pillow>=10.3.0
